imporve dev env
