+++ spec/item_spec.rb
@@ -1,7 +1,7 @@
 require 'spec_helper'
- 
+
 describe Item do
- 
+
   describe "#initialize" do
     it "is an Item" do
       new_item = Item.new("item_name","item_sell_in","item_quality")
